Каст данных к float

+++ src/models/model.py
@@ -144,7 +144,7 @@ def finetune_rubert(df, max_samples=None):
 
         def __getitem__(self, idx):
             item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
-            item["labels"] = torch.tensor(self.labels[idx])
+            item["labels"] = torch.tensor(self.labels[idx], dtype=torch.long)
             return item
 
         def __len__(self):
